+++ src/app/RackWidget.cpp
@@ -261,15 +261,27 @@ void RackWidget::cloneModule(ModuleWidget *m) {
 	json_t *moduleJ = m->toJson();
 	clonedModuleWidget->fromJson(moduleJ);
 	json_decref(moduleJ);
-	clonedModuleWidget->requestedPos = m->box.pos;
-	clonedModuleWidget->requested = true;
+	Rect clonedBox = clonedModuleWidget->box;
+	clonedBox.pos = m->box.pos;
+	requestModuleBoxNearest(clonedModuleWidget, clonedBox);
 	addModule(clonedModuleWidget);
 }
 
-void RackWidget::repositionModule(ModuleWidget *m) {
+bool RackWidget::requestModuleBox(ModuleWidget *m, Rect box) {
+	for (Widget *child2 : moduleContainer->children) {
+		if (m == child2) continue;
+		if (box.intersects(child2->box)) {
+			return false;
+		}
+	}
+	m->box = box;
+	return true;
+}
+
+bool RackWidget::requestModuleBoxNearest(ModuleWidget *m, Rect box) {
 	// Create possible positions
-	int x0 = roundf(m->requestedPos.x / RACK_GRID_WIDTH);
-	int y0 = roundf(m->requestedPos.y / RACK_GRID_HEIGHT);
+	int x0 = roundf(box.pos.x / RACK_GRID_WIDTH);
+	int y0 = roundf(box.pos.y / RACK_GRID_HEIGHT);
 	std::vector<Vec> positions;
 	for (int y = maxi(0, y0 - 4); y < y0 + 4; y++) {
 		for (int x = maxi(0, x0 - 200); x < x0 + 200; x++) {
@@ -278,27 +290,18 @@ void RackWidget::repositionModule(ModuleWidget *m) {
 	}
 
 	// Sort possible positions by distance to the requested position
-	Vec requestedPos = m->requestedPos;
-	std::sort(positions.begin(), positions.end(), [requestedPos](Vec a, Vec b) {
-		return a.minus(requestedPos).norm() < b.minus(requestedPos).norm();
+	std::sort(positions.begin(), positions.end(), [box](Vec a, Vec b) {
+		return a.minus(box.pos).norm() < b.minus(box.pos).norm();
 	});
 
 	// Find a position that does not collide
-	for (Vec pos : positions) {
-		Rect newBox = Rect(pos, m->box.size);
-		bool collides = false;
-		for (Widget *child2 : moduleContainer->children) {
-			if (m == child2) continue;
-			if (newBox.intersects(child2->box)) {
-				collides = true;
-				break;
-			}
-		}
-		if (collides) continue;
-
-		m->box.pos = pos;
-		break;
+	for (Vec position : positions) {
+		Rect newBox = box;
+		newBox.pos = position;
+		if (requestModuleBox(m, newBox))
+			return true;
 	}
+	return false;
 }
 
 void RackWidget::step() {
@@ -308,16 +311,6 @@ void RackWidget::step() {
 	Vec moduleSize = moduleContainer->getChildrenBoundingBox().getBottomRight();
 	// We assume that the size is reset by a parent before calling step(). Otherwise it will grow unbounded.
 	box.size = box.size.max(moduleSize);
-
-	// Reposition modules
-	for (Widget *child : moduleContainer->children) {
-		ModuleWidget *module = dynamic_cast<ModuleWidget*>(child);
-		assert(module);
-		if (module->requested) {
-			repositionModule(module);
-			module->requested = false;
-		}
-	}
 
 	// Autosave every 15 seconds
 	if (gGuiFrame % (60*15) == 0) {
@@ -344,9 +337,12 @@ struct AddModuleMenuItem : MenuItem {
 	Vec modulePos;
 	void onAction() {
 		ModuleWidget *moduleWidget = model->createModuleWidget();
-		moduleWidget->requestedPos = modulePos.minus(moduleWidget->box.getCenter());
-		moduleWidget->requested = true;
 		gRackWidget->moduleContainer->addChild(moduleWidget);
+		// Move module nearest to the mouse position
+		Rect box;
+		box.size = moduleWidget->box.size;
+		box.pos = modulePos.minus(box.getCenter());
+		gRackWidget->requestModuleBoxNearest(moduleWidget, box);
 	}
 };