+++ src/app/SVGPanel.cpp
@@ -6,11 +6,7 @@ namespace rack {
 
 struct PanelBorder : TransparentWidget {
 	void draw(NVGcontext *vg) {
-		nvgBeginPath(vg);
-		nvgRect(vg, 0.0, 0.0, box.size.x, box.size.y);
-
 		NVGcolor borderColor = nvgRGBAf(0.5, 0.5, 0.5, 0.5);
-
 		nvgBeginPath(vg);
 		nvgRect(vg, 0.5, 0.5, box.size.x - 1.0, box.size.y - 1.0);
 		nvgStrokeColor(vg, borderColor);
@@ -21,6 +17,8 @@ struct PanelBorder : TransparentWidget {
 
 
 void SVGPanel::setBackground(std::shared_ptr<SVG> svg) {
+	clearChildren();
+
 	SVGWidget *sw = new SVGWidget();
 	sw->wrap();
 	sw->svg = svg;