+++ src/core/core.hpp
@@ -21,3 +21,18 @@ struct MIDICCToCVWidget : ModuleWidget {
 	MIDICCToCVWidget();
 	void step();
 };
+
+struct BridgeWidget : ModuleWidget {
+	BridgeWidget();
+};
+
+struct BlankWidget : ModuleWidget {
+	Panel *panel;
+	Widget *topRightScrew;
+	Widget *bottomRightScrew;
+	Widget *rightHandle;
+	BlankWidget();
+	void step();
+	json_t *toJson();
+	void fromJson(json_t *rootJ);
+};